+++ parallel_pathfinding/parallel_pathfinding/CustomClasses/Graphing/Node/GridNode.cs
@@ -6,61 +6,30 @@ namespace parallel_pathfinding.CustomClasses.Graphing
 {
     public class GridNode : Node
     {
-        private bool _isWall = false;
-        public List<Tuple<GridNode, double>> Neighbours { get; }
-        private GridNode _parent;
-        private Tuple<int, int> Position;
+        public GridNode Parent { get; set; }
+        public int PositionX { get; }
+        public int PositionY { get; }
+        public bool IsWall { get; private set; } = false;
+
+        private double GCost { get; set; } = 0; //current lenght from starting Node, starting Node has g = 0
+        public double HCost { get; set; } = 0;
 
         public GridNode(int PositionX, int PositionY) : base()
         {
-            this.Neighbours = new List<Tuple<GridNode, double>>();
-            this.Position = new Tuple<int, int>(PositionX, PositionY);
+            this.PositionX = PositionX;
+            this.PositionY = PositionY;
         }
-
-        public void AddNeighbour(GridNode Neighbour, double Distance)
-        {
-            Neighbours.Add(new Tuple<GridNode, double>(Neighbour, Distance));
-        }
-
-        public int[] GetXY()
-        {
-            int[] coordinates = new int[2];
-            coordinates[0] = Position.Item1;
-            coordinates[1] = Position.Item2;
-            return coordinates;
-        }
-
-        public GridNode GetParent()
-        { return _parent; }
-
-        public void SetParent(GridNode parent)
-        { _parent = parent; }
 
         public void MakeWall()
         {
-            _isWall = true;
+            IsWall = true;
         }
 
         public void UnmakeWall()
         {
-            _isWall = false;
+            IsWall = false;
         }
 
-        public bool IsWall()
-        {
-            return _isWall;
-        }
-
-
+        public double FCost() => GCost + HCost; 
     }
-    /*
-    public class GridNode : Node
-    {
-        public Tuple<int, int> Coordinates { get; private set; }
-        public GridNode(int CoordinateX, int CoordinateY)
-        {
-            Coordinates = new Tuple<int, int>(CoordinateX, CoordinateY);
-        }
-    }
-    */
 }
